Chose JSON Processing as the JSON (de)serialization API
Reasoning can be found at #247.
